Size parameter of xduparray also needs to be uint32_t

+++ src/slurmctld/proc_req.h
@@ -74,7 +74,7 @@ extern int slurm_fail_job(uint32_t job_id);
 
 /* Copy an array of type char **, xmalloc() the array and xstrdup() the
  * strings in the array */
-extern char **xduparray(uint16_t size, char ** array);
+extern char **xduparray(uint32_t size, char ** array);
 
 #endif /* !_HAVE_PROC_REQ_H */
